authorization/authorization_code_python/requirements.txt
spotipy
requests
python-dotenv
Flask
urllib3
os
datetime